+++ packages/cli/doc-core/src/node/createBuilder.ts
@@ -34,7 +34,7 @@ async function createInternalBuildConfig(
   const themeDir = (await fs.pathExists(CUSTOM_THEME_DIR))
     ? CUSTOM_THEME_DIR
     : path.join(PACKAGE_ROOT, 'src', 'theme-default');
-
+  const checkDeadLinks = config.doc?.markdown?.checkDeadLinks ?? false;
   // Process doc config by plugins
   for (const plugin of docPlugins) {
     if (typeof plugin.config === 'function') {
@@ -149,6 +149,10 @@ async function createInternalBuildConfig(
             config: windiConfig,
           }),
         );
+        config.cache = {
+          // If checkDeadLinks is true, we should use memory cache to avoid skiping mdx-loader when starting dev server again
+          type: checkDeadLinks ? 'memory' : 'filesystem',
+        };
         return config;
       },
     },