+++ packages/generator/generator-plugin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/generator-plugin
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies [4568607]
+  - @modern-js/generator-common@3.0.2
+  - @modern-js/generator-utils@3.0.2
+  - @modern-js/new-action@2.0.2
+  - @modern-js/utils@2.0.2
+  - @modern-js/plugin-i18n@2.0.2
+
 ## 3.0.1
 
 ### Patch Changes