+++ packages/server/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/server-plugin
 
+## 2.0.2
+
+### Patch Changes
+
+- @modern-js/utils@2.0.2
+- @modern-js/plugin@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes